requirements.txt
Django==5.1.5
psycopg2_binary==2.9.10
djangorestframework==3.15.2
whitenoise==6.8.2
django-cors-headers==4.6.0
fontawesomefree==6.6.0
pillow==11.1.0
python-dotenv==1.0.1